requirements.txt
Flask==2.3.3
pandas==2.2.2
gunicorn==21.2.0
requests==2.31.0
google-generativeai==0.4.1
